+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.1.8 - Unreleased
 
 - `#[allow(clippy::all)]` in generated code
+- `is_empty` functions on maps
 
 # 0.1.7 - 2025-06-02